+++ crates/miden-objects/src/lib.rs
@@ -10,7 +10,14 @@ pub mod error;
 pub(crate) mod test_utils;
 
 pub use error::{ConversionError, ConversionResultExt};
-pub use miden_protobuf::{BuildUnchecked, DecodeMessage, Decoded, Verify, VerifyWith};
+pub use miden_protobuf::{
+    BuildUnchecked,
+    DecodeMessage,
+    DecodeMessageExt,
+    Decoded,
+    Verify,
+    VerifyWith,
+};
 pub use prost;
 
 /// Generated canonical Protobuf messages.