+++ crates/miden-protocol/src/lib.rs
@@ -68,6 +68,7 @@ pub mod utils;
 pub mod vm {
     pub use miden_assembly_syntax::ast::{AttributeSet, QualifiedProcedureName};
     pub use miden_core::advice::{AdviceInputs, AdviceMap, AdviceStack};
+    pub use miden_core::deferred::MAX_PRECOMPILE_ROOTS;
     pub use miden_core::events::{EventId, EventName, SystemEvent};
     pub use miden_core::program::{Program, ProgramInfo};
     pub use miden_mast_package::debug_info::{DebugSourceNodeId, PackageDebugInfo};